trying to change the way manual detection works

+++ recipes/deploy.rb
@@ -8,7 +8,6 @@ node["deploy"].each do |application, deploy|
     c.each do |app_name, app_config|
       Chef::Log.debug("Evaluating #{app_name}...")
       e = EnvHelper.new app_name, app_config, deploy, node
-      next if e.manual?
 
       image = app_config["image"]
       containers = app_config["containers"] || 1
@@ -29,22 +28,23 @@ node["deploy"].each do |application, deploy|
           end
         end
 
-        execute "kill running #{app_name}#{i} container" do
-          Chef::Log.info("Killing running #{application}/#{app_name}#{i} container...")
-          command "docker kill #{app_name}#{i}"
-          only_if "docker ps -f status=running | grep ' #{app_name}#{i} '"
-        end
+        unless e.manual?
+          execute "kill running #{app_name}#{i} container" do
+            Chef::Log.info("Killing running #{application}/#{app_name}#{i} container...")
+            command "docker kill #{app_name}#{i}"
+            only_if "docker ps -f status=running | grep ' #{app_name}#{i} '"
+          end
 
-        execute "remove stopped #{app_name}#{i} container" do
-          Chef::Log.info("Removing the #{application}/#{app_name}#{i} container...")
-          command "docker rm  #{app_name}#{i}"
-          only_if "docker ps -a | grep ' #{app_name}#{i} '"
-        end
-
-        execute "migrate #{app_name}#{i} container" do
-          Chef::Log.info("Migrating #{app_name}#{i}...")
-          command "docker run --rm #{e.env_string(environment)} #{e.links} #{e.volumes} #{e.volumes_from} #{image} #{app_config["migration"]}"
-          only_if { e.migrate? && i == 0}
+          execute "remove stopped #{app_name}#{i} container" do
+            Chef::Log.info("Removing the #{application}/#{app_name}#{i} container...")
+            command "docker rm  #{app_name}#{i}"
+            only_if "docker ps -a | grep ' #{app_name}#{i} '"
+          end
+          execute "migrate #{app_name}#{i} container" do
+            Chef::Log.info("Migrating #{app_name}#{i}...")
+            command "docker run --rm #{e.env_string(environment)} #{e.links} #{e.volumes} #{e.volumes_from} #{image} #{app_config["migration"]}"
+            only_if { e.migrate? && i == 0}
+          end
         end
 
         execute "launch #{app_name}#{i} container" do
@@ -52,7 +52,7 @@ node["deploy"].each do |application, deploy|
           environment["RELEASE_TAG"] = history.stdout.strip
           Chef::Log.info("Launching #{image}...")
           command "docker run -d -h #{e.hostname i} --name #{app_name}#{i} #{e.ports} #{e.env_string(environment)} #{e.links} #{e.volumes} #{e.volumes_from} #{e.entrypoint} #{image} #{e.cmd i}"
-          only_if { e.auto? }
+          not_if "docker ps -f status=running | grep ' #{app_name}#{i} '"
         end
 
         cron "#{app_name}#{i} cron" do
